backend/careminder/requirements.txt
django
djangorestframework-simplejwt
django-cors-headers
django-filter